:recycle: Refactor styles files
