+++ mudiaga-uwojeya/sass/base/animations.sass
@@ -1,11 +1,8 @@
-@keyframes slide-up {
-    from {
-        opacity: 0;
-        transform: translateY(3rem);
-    }
+@keyframes slide-up 
+    from 
+        opacity: 0
+        transform: translateY(3rem)
 
-    to {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
+    to 
+        opacity: 1
+        transform: translateY(0)
